Append version info to build files
This is to ensure that our CDN doesn't serve up old versions of any built files. Library files have version numbers in the file name, and we'll ignore images and fonts since they're unlikely to change.

This has the unfortunate effect of placing a new built file (index.html) outside of the build directory, but we'll punt on making that more elegant.