Create bare-bones api caller service
Seems like everything that must interact with the api must know to do
many different things, such as having the correct headers, including
the config's apiHost, etc.

This causes lots of little issues down the road, the biggest being if
you have to do a special thing to one call or just want to make a
blanket change to every call.  This service solves this by having
providing methods that automatically append the correct url and
headers.

There is plenty of other things this can do, but as the commit says,
this is just a barebones version.