Add getAll to entity service
Now that we have a service for handling api calls, might as well start
putting the actual endpoint calls in the services they are related
to.

Started with the getAll entities endpoint and converted the homeCtrl
call to test if it works. Next up will be converting the other usages
of getAll to the services, then move the rest of the api calls to
their respective services