Move all api-related services to api directory
This is a followup to the previous commit were I began to move api
calls for all entities to the entityService.  In this commit, took
that to the next level and moved all calls to their respective
service, along with moving all services to exist in the same api
directive under the new angular module.

On top of this, cleaned up the code a bit for the files that I had to
work on.  Looking at what I changed, I realize that I have set it up
in a way that would be convienent for ES6 import/export, but not for
the current version + angular.  Next step is to build a util service
for storing this methods there and then have everyone use the util
service until this project can be converted to using
babel/webpack/something for handling import/export