Add cgUtilService to provide utility functions between controllers
Follow up to previous commit, where I noticed that while "cleaning
up",  I had accidentally caused more duplicate code.  This simply adds
a utils service that should help that a bit.

The first few functions that I put in there are the isDefined function
and the couple we use for the Models in the API services.  That seemed
like that quickest, easiest win.  More could be added as duplicates
are found.

Also, this is a general catch-all utils service.  Would be a good idea
to instead split into multiple util services as the amount of
functions grows.  Say, a is* service for isDefined, isObject,
etc. that is separate from the util methods used for building models.