Remove many scope variables that are no longer used
This was a fun one.  Had to dig around in all the pages that inherit
scope from HomeCtrl to figure this out, but there is quite a few
variables set in HomeCtrl's scope that just aren't used anymore or are
not actually doing anything.

A big example is clickedLocation.  I'm guessing this was a feature or
at least planned at one point, but now is just set at the beginning,
never updated, and then checked in quite a few spots in the network
graph.   Got rid of the checks and everything works the same as before
and then just got rid of the original assignment, as it was just a
null assignment.