Clean up edit entity and detail/overview
They don't seem related and that is mainly because they aren't.  The
edit entity clean up was mainly just attempting to remove as much as
possible from scope.  While attempting that, found a weird thing on
how the overview was being loaded, so went ahead and cleaned that
while I was there as well.