Nasty commit in attempt to untangle network directive
Didn't realize what I had gotten into before stepping into this
tangled mess.  It seems that network directive is DEEPLY connected
with the rest of the app without any obvious connections.

This is merely a first pass attempt and does not cover the entire
thing.  I created a new service specifically for managing the state of
the network, such as what sizeBy is, minConnections, the connection
and entity data, etc.  This should help give a single source location
for talking with the network directive.  New data?  Send to Service.
Need to know what it thinks is happening?  Ask the service.  With the
design, only allows a single state for the network, but that is fine
as I don't think we plan on multiple graphs at once.  Would be neat
later on I suppose, not sure how it would look though.

Next steps is to really dig down more on building services for
managing state of the app.  Maybe a main service for storing the api
data and current selected data, and the network pulling from that.
Need something, as the map is still a whole other beast entirely.