Create Network Graph Directive and main service for app
Decided to try both of the things I mentioned in the previous commit
and they worked out well enough.  First was to create a internal
directive just for the d3 part of the network.  The second was
creating a main service for the app to share state about the app.

For the d3 directive, this made the code quite a bit cleaner, along
with giving us more flexibility on watching changes to things.
Haven't converted everything 100%, but progress.

As for the main service, the only thing it does so far is keep up with
what is the current entity and has a method for checking if we are
currently in mobile mode.  This simplifies quite a bit, but not
everything.  Still have the issue of most things pulling the entity
list from HomeCtrl, but that is as simple as adding the entity list to
main service then converting to either watchers or event listeners.
Haven't decided with I like better.