Separate app dependency on homeCtrl for entity list
This was the biggest one that I could do.  Every part of the app
expected scope to contain a property called `entities` that was
provided by homeCtrl.  This removes that dependency by moving the
retrieval of the entity list to the main service.

This allowed for quite a bit of simplification, the biggest in the
network directive.  After everything was done, the only thing that the
controller is doing is managing the loader bar.  I think the next part
will be to have the map contained to a directive like I did for
network, then have the network controller become the central section
controller so that it controls if the network or map is shown.  This
will clean up index a bit and give us a single point to fix CSS
problems, as everything in the center of the page will be contained in
that one directive/controller.