Move Map to directive managed by new central component
The old top-level network directive is now actually the component that
handles the switching between map and network.  Also will move the
search bar into it as well, so that everything that goes in the middle
column will be handled by a single component.

I didn't really change Map directive too much, mainly just removing
the last bit of the dependencies on parent state that was left and
then added a listener to reload when data arrives, that way if it is
the first one to show on the page, it will load the data properly.