Moving CurrentEntity management to a more event-based system
Full Disclosure: I am not 100% sure what the CSS parts of this commit
are doing.  As can be seen from the timestamps between this commit and
the previous one, I haven't worked on this in close to 6 months.  Most
of that was already in place when I opened the project up.  From what
I could tell, I was working on standardizing the layout of the app and
moving each directive into the correct section.  Current me approves
of past me's decisions in this case, but I wasn't focused on that
tonight, so just committing what he had worked on, as it does not
actually break anything.

As for the real meat of this commit, while working on the issue of the
network not seeing the updated current entity, I noticed a very bad
sign of how I was setting things up.  I put a print out in the getter
method for current entity and saw that it was being fired constantly.
Not ideal for what I want to do, so decided to move it to an event
system instead of a watcher-based system.

What this means practically is that instead of watching for changes to
the current entity, you can instead subscribe to the event that will
be broadcast from $rootScope from the main service whenever current
entity is updated.  This reduces total calls to the page and is
probably the way it should have worked to begin with.  A part of me
cringed using the $rootScope, as I try to avoid "hacky" things like
that, but honestly it was a global variable before, so doesn't change
the scope of it now, simply reduces total function calls in the app.