Fix links in network not showing and increase default minConnection
Links not showing was a simple fix, I was accessing the API response
incorrectly.  Everything else worked fine.  After that, noticed that
it was significantly slower, so tried speeding up by removing
redundant connections.  Didn't help, so just set the default
minConnection to a higher value and will look into it more later.

Also noticed, which might contribute to the problem above, that there
seems to be more connections drawn now than in the original code.
This was what lead me to believe it was the redundent connections, but
checking the original code did not show any filters like that.  In
fact, no filters were applied to the links, so I am unsure why it
seems there are more now.  Might be my imagination.