Create right pane directive and add start/stop edit event
Quite pleased with myself on how easily that went, partially as it is
currently one in the morning.  Followed up with what I assume I had
started back in August before I abandoned this project and wrapped
everything that appears on the right pane into a single directive.
This will make styling easier, but didn't actually fix styling, so the
CSS is still ugly.

Also, while testing that out, noticed that the Create/Edit pane didn't
work properly with the new current entity system, so updated that to
not only update itself based on the current entity, but also setup an
event for starting and stopping create/edit.  This won't be perfect,
as I have not tested every part of the app with this new event, but
seems to mesh well with the graph, details, and overview, which are
the major sections right now.